Moved consumer sessions to /consumer url and dashboard to home page
