+++ src/vng/templates/components/navigation/navigation-lateral.html
@@ -4,7 +4,7 @@
     <ul class="nav">
         {% if user.is_authenticated %}
         <li class="nav-item">
-	  <a class="nav-link" href="{% url 'testsession:dashboard' %}">
+	  <a class="nav-link" href="{% url 'dashboard' %}">
 	    <i class="nav-icon icon-speedometer"></i>{% trans "Dashboard" %}</a>
 	</li>
