+++ src/vng/urls.py
@@ -1,10 +1,12 @@
 from django.conf import settings
 from django.conf.urls import include, url
+from django.urls import path
 from django.contrib import admin
 from django.contrib.auth import views as auth_views
 from django.views.generic.base import TemplateView
 from .decorators import anonymous_required
 from .base_url import *
+from . import views
 
 urlpatterns = base_urlpatterns + [
     # redirect the request to the testession
@@ -14,5 +16,6 @@ urlpatterns = base_urlpatterns + [
     url(r'^api/v1/', include('vng.servervalidation.urls_api', namespace='apiv1server')),
     url(r'^api/v1/', include('vng.openApiInspector.urls_api', namespace='apiv1inspector')),
     url(r'^server/', include('vng.servervalidation.urls', namespace='server_run')),
-    url(r'^', include('vng.testsession.urls', namespace='testsession')),
+    url(r'^consumer/', include('vng.testsession.urls', namespace='testsession')),
+    path('', views.Dashboard.as_view(), name='dashboard'),
 ]
